Combine omnipotion slice into infusions
